ARM: dts: OMAP5: Add ocp address space for all hwmods
Patch adds the OCP address space for all missing hwmod from existing
DT file. Note that the 'simple-bus' compatible is and added and *state*
is kept as 'disable' by purpose to ensure that for these hwmod, devices
are not getting created.

While adapting device drivers for DT, the respective devices compatible
and status needs to be updated. For nowi, added dt device data is used
only for initializing the IPs in hwmod core layer.

Cc: Benoit Cousson <b-cousson@ti.com>

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>